Extract line length tests to their own test files
