+++ Tests/SwiftLintFrameworkTests/RulesTests.swift
@@ -10,7 +10,6 @@ import SwiftLintFramework
 import XCTest
 
 // swiftlint:disable file_length
-// swiftlint:disable:next type_body_length
 class RulesTests: XCTestCase {
 
     func testClassDelegateProtocol() {
@@ -132,29 +131,6 @@ class RulesTests: XCTestCase {
 
     func testLegacyConstructor() {
         verifyRule(LegacyConstructorRule.description)
-    }
-
-    func testLineLength() {
-        verifyRule(LineLengthRule.description, commentDoesntViolate: false, stringDoesntViolate: false)
-
-        // Perform additional tests with the ignores_urls settings enabled
-        let url = "https://github.com/realm/SwiftLint"
-        let triggeringLines = [String(repeating: "/", count: 121) + "\(url)\n"]
-        let nonTriggeringLines = ["\(url) " + String(repeating: "/", count: 118) + " \(url)\n",
-            "\(url)/" + String(repeating: "a", count: 120)]
-
-        let baseDescription = LineLengthRule.description
-        let nonTriggeringExamples = baseDescription.nonTriggeringExamples + nonTriggeringLines
-        let triggeringExamples = baseDescription.triggeringExamples + triggeringLines
-        let description = RuleDescription(identifier: baseDescription.identifier,
-                                          name: baseDescription.name,
-                                          description: baseDescription.description,
-                                          nonTriggeringExamples: nonTriggeringExamples,
-                                          triggeringExamples: triggeringExamples,
-                                          corrections: baseDescription.corrections)
-
-        verifyRule(description, ruleConfiguration: ["ignores_urls": true],
-                   commentDoesntViolate: false, stringDoesntViolate: false)
     }
 
     func testMark() {
@@ -389,7 +365,6 @@ extension RulesTests {
             ("testLegacyNSGeometryFunctions", testLegacyNSGeometryFunctions),
             ("testLegacyConstant", testLegacyConstant),
             ("testLegacyConstructor", testLegacyConstructor),
-            ("testLineLength", testLineLength),
             ("testMark", testMark),
             ("testNesting", testNesting),
             ("testNimbleOperator", testNimbleOperator),
